requirements.txt
# Core Framework & Server
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
websockets==15.0.1
watchfiles==1.1.0

# Database & Migration
SQLAlchemy==2.0.43
alembic==1.16.5
Mako==1.3.10
psycopg2-binary==2.9.10  # PostgreSQL DB 드라이버
greenlet==3.2.4

# Security & Authentication
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# Environment & Config
python-dotenv==1.1.1
PyYAML==6.0.2

# Data Validation & Utilities
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1

onnxruntime
opencv-python-headless
numpy
pillow
shapely
rapidfuzz
pytesseract
python-multipart

# Testing (Test Runners & Dependencies)
pytest==8.4.1
pluggy==1.6.0
iniconfig==2.1.0
click==8.2.1
colorama==0.4.6
Pygments==2.19.2

# Networking & Certs
certifi==2025.8.3
idna==3.10
httpx==0.28.1